requirements.txt
git+https://github.com/openai/whisper.git 
ffmpeg == 1.4
openai == 0.27.8
python-dotenv == 1.0.0
